requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
weaviate-client==3.19.0
openai==0.28.0
docx2txt==0.8
tiktoken==0.5.0
python-docx~=1.1.2